Support DryRun attribute (and fix initial status)
Hooking a new deployment to WPA might be hazardous:
will this drop all my pods all of a sudden? hard to anticipate.

Temporarily setting very high and very low watermarks won't help
much to guess what's the outcome would be with normal values.

This might not be material for an upstream version, but by
then we'll hopefully have gained experience and confidence
hooking correct values.

While at it: On fresh new WPAs, `Status.CurrentMetrics` obtained
by DeepCopy is unserialized as nil (not an empty slice), which
prevents the WPA status merge and update to work, so the WPA
status remains empty and we log errors in loop:

```
2019-09-17T18:30:58.399Z        ERROR   wpa_controller  Error during reconcileWPA       {"Request.Namespace": "datadog", "Request.Name": "tdeploy", "error": "WatermarkPodAutoscaler.datadoghq.com \"tdeploy\" is invalid: status: Invalid value: map[string]interface {}{\"conditions\":[]interface {}{map[string]interface {}{\"type\":\"AbleToScale\", \"status\":\"True\", \"lastTransitionTime\":\"2019-09-17T18:30:58Z\", \"reason\":\"SucceededRescale\", \"message\":\"the HPA controller was able to update the target scale to 4\"}}, \"lastScaleTime\":\"2019-09-17T18:30:58Z\", \"currentReplicas\":1, \"desiredReplicas\":4, \"currentMetrics\":interface {}(nil)}: validation failure list:\ncurrentMetrics in body must be of type array: \"null\""}
```