Fix logs-attributes
the call to the function `GetLogAttrsFromWpa` was made before getting
the data inside the `WatermarkPodAutoscaler` struct instance.

This PR move the call to the function after the `client.Get(...)` that
populates the `instance` with the data.
Also rework a bit the function `GetLogAttrsFromWpa` to return an error
in case of invalid annotation value.