fix: allow wpa status update even if dry-run:true (#120)
In specific configuration if the status is empty (new WPA) in dry
run mode, and if `currentReplica` > max or < min, the status is not
created properly because wpa.status.CurrentMetrics is a empty slice
but not a nil slice.

This commit remove the initialisation of `metricStatuses` to an empty
slice. like this the status should be created.