fix: improve how we filter pods from other deployments
previously the way we computed the list of pod ready corresponding
to a WPA was not enough precise. In some case if 2 deployments in
the same Namespace was using the same Selector can trigger a bug
in the wpa where both deployment's pods are taken into account when
calculating the current replicas value.