recommender: Initial support for third-party replica recommenders (#223)
* recommender: Initial support for third-party replica recommenders

This is a first step to implement the interface defined in https://github.com/DataDog/agent-payload/pull/348

Currently this simply adds a non-implemented recommender client and the necesary logic to handle
the results. It does not allow the usage all the features made available by the recommendation
API and keep it to the minimum required to make use of it (for example simple TargetType).

Simple tests have been implemented and this was implemented to re-use as much code as possible in the replica_calculator.

Follow-up PRs will actually implement the recommender client and add the necessary tests and documentation.

* recommender: finish the initial implementation

* lint

* Restore manifests

* Move the mock to a test file