Supports setting wpa default TLS config for the recommender API
For environment where certificates can be the same for all WPA resources,
it might be cumbersome to specify the same client certificate in all
the recommender specs.
By using the `--tls-*` option it is possible to set default CA, certificate
and key files (along with a few other TLS options) that will be used
when the recommender specs don't contain such information.