Retry if certificate & key loaded don't match
For environments where client certificate loaded from disks are renewed
it is very possible that the certificate generator can't write both files
atomically. More care should be used when reading so that if the certificate
and the key don't match, reading is retried until there's either a match or
a failure.