Add TLS configuration to Recommender http client (#252)
* Add TLS configuration to Recommender http client

This allows to set client certificate along with server CA allowing
the recommender server to implement mTLS and validate client certificate.
Note that this will work only with a default client implementing an
http.Transport. Any other RoundTripper will need to implement their
own TLS implementation.
The implementation here reloads the certificate and private key for
every request to a given recommender endpoint. This is to support
systems where certificate can be rotated from time to time.

* Implement certificate caching

To prevent reloading the certificate and key from disk at every new
HTTPs connection we cache the certificate in a central cache (up until
it is either expired or too old). The cache is trimmed every 10 minutes
to remove stale entries (not accessed after 10 min).

* Supports setting wpa default TLS config for the recommender API

For environment where certificates can be the same for all WPA resources,
it might be cumbersome to specify the same client certificate in all
the recommender specs.
By using the `--tls-*` option it is possible to set default CA, certificate
and key files (along with a few other TLS options) that will be used
when the recommender specs don't contain such information.

* Retry if certificate & key loaded don't match

For environments where client certificate loaded from disks are renewed
it is very possible that the certificate generator can't write both files
atomically. More care should be used when reading so that if the certificate
and the key don't match, reading is retried until there's either a match or
a failure.

* Declare the recommender tls_config in the CRD

* Refactor tests for a bit more readability

* Review fixes

* Remove LRU mentions as the cache is not an LRU cache
* Cache errors for 1 min to prevent busy looping reading invalid certs

* Review fix: reorder TLS functions

So that types and methods are close together.

* Regenerate all bundle and manifests