doc(README): Update docs to explain why you shouldn't set replicas in the deployment (#287)
* doc(README): Update docs to explain why you shouldn't set replicas in the deployment

Problem

If you set the `replicas` value on your deployment, then every time you apply the deployment, the number of replicas will be reset to that value.
This will cause a race between the Kubernetes controller and WPA in setting the number of replicas and fluctuate the number down and up when you apply the deployment.

Solution

Add a note to the README explaining why you shouldn't set replicas in the deployment, and the issues you may see if you do.

Result

Users will be less likely to set replicas in their deployment and will avoid the issues described above.

* fix(README.md): include link to K8s docs for HPA which are relevant to the discussion of replicas